Reorganize messaging code into correct packages
