+++ src/main/kotlin/com/appifyhub/monolith/controller/creator/MessageTemplateController.kt
@@ -1,19 +1,19 @@
-package com.appifyhub.monolith.controller.messaging
+package com.appifyhub.monolith.controller.creator
 
 import com.appifyhub.monolith.controller.common.Endpoints
-import com.appifyhub.monolith.domain.messaging.MessageTemplate
-import com.appifyhub.monolith.domain.messaging.Variable
+import com.appifyhub.monolith.domain.creator.messaging.MessageTemplate
+import com.appifyhub.monolith.domain.creator.messaging.Variable
 import com.appifyhub.monolith.domain.user.UserId
 import com.appifyhub.monolith.network.common.SimpleResponse
 import com.appifyhub.monolith.network.mapper.toDomain
 import com.appifyhub.monolith.network.mapper.toNetwork
-import com.appifyhub.monolith.network.messaging.MessageResponse
-import com.appifyhub.monolith.network.messaging.MessageTemplateResponse
-import com.appifyhub.monolith.network.messaging.VariableResponse
-import com.appifyhub.monolith.network.messaging.ops.DetectVariablesRequest
-import com.appifyhub.monolith.network.messaging.ops.MessageInputsRequest
-import com.appifyhub.monolith.network.messaging.ops.MessageTemplateCreateRequest
-import com.appifyhub.monolith.network.messaging.ops.MessageTemplateUpdateRequest
+import com.appifyhub.monolith.network.creator.messaging.MessageResponse
+import com.appifyhub.monolith.network.creator.messaging.MessageTemplateResponse
+import com.appifyhub.monolith.network.creator.messaging.VariableResponse
+import com.appifyhub.monolith.network.creator.messaging.ops.DetectVariablesRequest
+import com.appifyhub.monolith.network.creator.messaging.ops.MessageInputsRequest
+import com.appifyhub.monolith.network.creator.messaging.ops.MessageTemplateCreateRequest
+import com.appifyhub.monolith.network.creator.messaging.ops.MessageTemplateUpdateRequest
 import com.appifyhub.monolith.service.access.AccessManager
 import com.appifyhub.monolith.service.access.AccessManager.Privilege
 import com.appifyhub.monolith.service.messaging.MessageTemplateService
