+++ src/main/kotlin/com/appifyhub/monolith/network/mapper/MessageTemplateMapper.kt
@@ -1,17 +1,17 @@
 package com.appifyhub.monolith.network.mapper
 
-import com.appifyhub.monolith.domain.messaging.Message
-import com.appifyhub.monolith.domain.messaging.MessageTemplate
-import com.appifyhub.monolith.domain.messaging.Variable
-import com.appifyhub.monolith.domain.messaging.ops.MessageTemplateCreator
-import com.appifyhub.monolith.domain.messaging.ops.MessageTemplateUpdater
+import com.appifyhub.monolith.domain.creator.messaging.Message
+import com.appifyhub.monolith.domain.creator.messaging.MessageTemplate
+import com.appifyhub.monolith.domain.creator.messaging.Variable
+import com.appifyhub.monolith.domain.creator.messaging.ops.MessageTemplateCreator
+import com.appifyhub.monolith.domain.creator.messaging.ops.MessageTemplateUpdater
 import com.appifyhub.monolith.domain.user.UserId
-import com.appifyhub.monolith.network.messaging.MessageResponse
-import com.appifyhub.monolith.network.messaging.MessageTemplateResponse
-import com.appifyhub.monolith.network.messaging.VariableResponse
-import com.appifyhub.monolith.network.messaging.ops.MessageInputsRequest
-import com.appifyhub.monolith.network.messaging.ops.MessageTemplateCreateRequest
-import com.appifyhub.monolith.network.messaging.ops.MessageTemplateUpdateRequest
+import com.appifyhub.monolith.network.creator.messaging.MessageResponse
+import com.appifyhub.monolith.network.creator.messaging.MessageTemplateResponse
+import com.appifyhub.monolith.network.creator.messaging.VariableResponse
+import com.appifyhub.monolith.network.creator.messaging.ops.MessageInputsRequest
+import com.appifyhub.monolith.network.creator.messaging.ops.MessageTemplateCreateRequest
+import com.appifyhub.monolith.network.creator.messaging.ops.MessageTemplateUpdateRequest
 import com.appifyhub.monolith.network.user.DateTimeMapper
 import com.appifyhub.monolith.service.messaging.MessageTemplateService.Inputs
