+++ src/test/kotlin/com/appifyhub/monolith/domain/mapper/MessageTemplateMapperTest.kt
@@ -3,7 +3,7 @@ package com.appifyhub.monolith.domain.mapper
 import assertk.assertThat
 import assertk.assertions.isDataClassEqualTo
 import assertk.assertions.isEqualTo
-import com.appifyhub.monolith.domain.messaging.ops.MessageTemplateUpdater
+import com.appifyhub.monolith.domain.creator.messaging.ops.MessageTemplateUpdater
 import com.appifyhub.monolith.storage.model.messaging.MessageTemplateDbm
 import com.appifyhub.monolith.util.Stubs
 import com.appifyhub.monolith.util.TimeProviderFake
