+++ src/test/kotlin/com/appifyhub/monolith/util/Stubs.kt
@@ -11,12 +11,12 @@ import com.appifyhub.monolith.domain.geo.Geolocation
 import com.appifyhub.monolith.domain.integrations.FirebaseConfig
 import com.appifyhub.monolith.domain.integrations.MailgunConfig
 import com.appifyhub.monolith.domain.integrations.TwilioConfig
-import com.appifyhub.monolith.domain.messaging.Message
-import com.appifyhub.monolith.domain.messaging.MessageTemplate
+import com.appifyhub.monolith.domain.creator.messaging.Message
+import com.appifyhub.monolith.domain.creator.messaging.MessageTemplate
 import com.appifyhub.monolith.domain.messaging.PushDevice
-import com.appifyhub.monolith.domain.messaging.Variable
-import com.appifyhub.monolith.domain.messaging.ops.MessageTemplateCreator
-import com.appifyhub.monolith.domain.messaging.ops.MessageTemplateUpdater
+import com.appifyhub.monolith.domain.creator.messaging.Variable
+import com.appifyhub.monolith.domain.creator.messaging.ops.MessageTemplateCreator
+import com.appifyhub.monolith.domain.creator.messaging.ops.MessageTemplateUpdater
 import com.appifyhub.monolith.domain.schema.Schema
 import com.appifyhub.monolith.domain.user.Organization
 import com.appifyhub.monolith.domain.user.SignupCode
@@ -39,15 +39,15 @@ import com.appifyhub.monolith.network.creator.user.ops.CreatorSignupRequest
 import com.appifyhub.monolith.network.integrations.FirebaseConfigDto
 import com.appifyhub.monolith.network.integrations.MailgunConfigDto
 import com.appifyhub.monolith.network.integrations.TwilioConfigDto
-import com.appifyhub.monolith.network.messaging.MessageResponse
-import com.appifyhub.monolith.network.messaging.MessageTemplateResponse
+import com.appifyhub.monolith.network.creator.messaging.MessageResponse
+import com.appifyhub.monolith.network.creator.messaging.MessageTemplateResponse
 import com.appifyhub.monolith.network.messaging.PushDeviceResponse
 import com.appifyhub.monolith.network.messaging.PushDevicesResponse
-import com.appifyhub.monolith.network.messaging.VariableResponse
-import com.appifyhub.monolith.network.messaging.ops.MessageInputsRequest
+import com.appifyhub.monolith.network.creator.messaging.VariableResponse
+import com.appifyhub.monolith.network.creator.messaging.ops.MessageInputsRequest
 import com.appifyhub.monolith.network.messaging.ops.MessageSendRequest
-import com.appifyhub.monolith.network.messaging.ops.MessageTemplateCreateRequest
-import com.appifyhub.monolith.network.messaging.ops.MessageTemplateUpdateRequest
+import com.appifyhub.monolith.network.creator.messaging.ops.MessageTemplateCreateRequest
+import com.appifyhub.monolith.network.creator.messaging.ops.MessageTemplateUpdateRequest
 import com.appifyhub.monolith.network.messaging.ops.PushDeviceRequest
 import com.appifyhub.monolith.network.user.OrganizationDto
 import com.appifyhub.monolith.network.user.SignupCodeResponse
